Share a single E2E setup across fixtures
This PR consolidates the per-fixture E2E setup files into a single shared `test/e2e/_setup.ts`. The System, Token and Memo tests now use one composed LiteSVM client whose program plugins come from their respective generated fixtures, which additionally proves that plugins generated from independent Codama clients compose correctly on a single Kit client. The token helpers and the nonce instruction plan helper move into the shared file unchanged, and the Memo test's inline client creation is replaced by the shared factory. One semantic note: token tests previously drove `client.system` through the token fixture's bundled system client and now use the system fixture's, which is behaviourally identical. The `node-esm` fixture and the `dummy`/`anchor` tests are untouched.
