+++ test/e2e/_setup.ts
@@ -1,18 +1,37 @@
-import { type Address, type TransactionSigner, createClient, generateKeyPairSigner, lamports } from '@solana/kit';
+import {
+    type Address,
+    type ClientWithPayer,
+    type ClientWithRpc,
+    type GetMinimumBalanceForRentExemptionApi,
+    type TransactionSigner,
+    createClient,
+    generateKeyPairSigner,
+    lamports,
+    sequentialInstructionPlan,
+} from '@solana/kit';
 import { litesvm } from '@solana/kit-plugin-litesvm';
 import { airdropSigner, generatedSigner } from '@solana/kit-plugin-signer';
 
+import { memoProgram } from './memo/src/index.js';
+import {
+    SYSTEM_PROGRAM_ADDRESS,
+    getCreateAccountInstruction,
+    getInitializeNonceAccountInstruction,
+    getNonceSize,
+    systemProgram,
+} from './system/src/index.js';
 import {
     TOKEN_PROGRAM_ADDRESS,
     associatedTokenProgram,
     getMintSize,
     getTokenSize,
-    systemProgram,
     tokenProgram,
-} from '../src/index.js';
+} from './token/src/index.js';
 
 /**
- * Creates a funded LiteSVM client with the generated token-related program plugins.
+ * Creates a funded LiteSVM client composing the generated program plugins of
+ * every fixture. Since each plugin comes from an independently generated
+ * client, this also proves that generated plugins compose on a single client.
  *
  * @return A promise resolving to the configured test client.
  */
@@ -23,16 +42,48 @@ export const createTestClient = () => {
         .use(airdropSigner(lamports(1_000_000_000n)))
         .use(systemProgram())
         .use(tokenProgram())
-        .use(associatedTokenProgram());
+        .use(associatedTokenProgram())
+        .use(memoProgram());
 };
 
-/** A LiteSVM client configured for Token Program E2E tests. */
+/** A LiteSVM client configured with the generated program plugins of every fixture. */
 export type TestClient = Awaited<ReturnType<typeof createTestClient>>;
+
+/**
+ * Creates an instruction plan that creates and initialises a durable nonce account.
+ *
+ * @param client - The client used to determine the rent-exempt balance.
+ * @param nonce - The signer for the nonce account.
+ * @param nonceAuthority - The authority allowed to advance the nonce.
+ * @return A promise resolving to the sequential instruction plan.
+ */
+export const getCreateNonceInstructionPlan = async (
+    client: ClientWithPayer & ClientWithRpc<GetMinimumBalanceForRentExemptionApi>,
+    nonce: TransactionSigner,
+    nonceAuthority: TransactionSigner,
+) => {
+    const space = BigInt(getNonceSize());
+    const rent = await client.rpc.getMinimumBalanceForRentExemption(space).send();
+
+    return sequentialInstructionPlan([
+        getCreateAccountInstruction({
+            payer: client.payer,
+            newAccount: nonce,
+            lamports: rent,
+            space,
+            programAddress: SYSTEM_PROGRAM_ADDRESS,
+        }),
+        getInitializeNonceAccountInstruction({
+            nonceAccount: nonce.address,
+            nonceAuthority: nonceAuthority.address,
+        }),
+    ]);
+};
 
 /**
  * Creates and initialises a token mint.
  *
- * @param client - The configured Token Program test client.
+ * @param client - The configured test client.
  * @param mintAuthority - The authority allowed to mint tokens.
  * @param decimals - The number of decimal places used by the mint.
  * @return The address of the created mint.
@@ -66,7 +117,7 @@ export const createMint = async (
 /**
  * Creates and initialises a token account.
  *
- * @param client - The configured Token Program test client.
+ * @param client - The configured test client.
  * @param mint - The mint associated with the token account.
  * @param owner - The owner of the token account.
  * @return The address of the created token account.
@@ -96,7 +147,7 @@ export const createToken = async (client: TestClient, mint: Address, owner: Addr
 /**
  * Creates a token account and mints an initial amount into it.
  *
- * @param client - The configured Token Program test client.
+ * @param client - The configured test client.
  * @param mintAuthority - The authority allowed to mint tokens.
  * @param mint - The mint associated with the token account.
  * @param owner - The owner of the token account.
