+++ test/e2e/system/test/createAccount.test.ts
@@ -2,7 +2,7 @@ import { generateKeyPairSigner } from '@solana/kit';
 import { expect, test } from 'vitest';
 
 import { SYSTEM_PROGRAM_ADDRESS } from '../src/index.js';
-import { createTestClient } from './_setup.js';
+import { createTestClient } from '../../_setup.js';
 
 test('it creates a new empty account', async () => {
     // Given a newly generated account keypair and 42 bytes of space.
