+++ test/e2e/system/test/initializeNonceAccount.test.ts
@@ -2,7 +2,7 @@ import { generateKeyPairSigner } from '@solana/kit';
 import { expect, test } from 'vitest';
 
 import { NonceState, NonceVersion } from '../src/index.js';
-import { createTestClient, getCreateNonceInstructionPlan } from './_setup.js';
+import { createTestClient, getCreateNonceInstructionPlan } from '../../_setup.js';
 
 test('it creates and initialises a durable nonce account', async () => {
     // Given a nonce account and authority.
