+++ test/e2e/token/test/createAssociatedTokenIdempotent.test.ts
@@ -2,7 +2,7 @@ import { generateKeyPairSigner } from '@solana/kit';
 import { expect, test } from 'vitest';
 
 import { TOKEN_PROGRAM_ADDRESS } from '../src/index.js';
-import { createMint, createTestClient } from './_setup.js';
+import { createMint, createTestClient } from '../../_setup.js';
 
 test('it creates an associated token account idempotently', async () => {
     // Given a mint, its authority and a token owner.
